Print a reslr output object which is created by the reslr_mcmc function.

This will be very high level printing that the user can use to obtain information about the MCMC run using JAGS. The number of iterations and chains used by the user is printed In addition, the type of statistical model is printed.

# S3 method for reslr_output
print(x, ...)

Arguments

x

An object of class reslr_output

...

Other arguments (not supported)

Value

Returns high level information about the reslr_output object, i.e. the number of iterations and chains used.

Examples

data <- NAACproxydata %>% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
jags_output <- reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t")
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 208
#>    Unobserved stochastic nodes: 107
#>    Total graph size: 1114
#> 
#> Initializing model
#> 
print(x = jags_output)
#> This is a valid reslr output object with 104 observations and  1 site(s).
#> There are  1  proxy site(s) and  0  tide gauge site(s).
#> The age units are; Common Era. 
#> The model used was the Errors-in-Variables Simple Linear Regression model.
#> The input data has been run via reslr_mcmc and has produced 3000 iterations over 3 MCMC chains.