Produces summaries and convergence diagnostics for an object created with reslr_mcmc.

A warning message will appear if the model has not been converge. If this appears the user is recommended to re-run the model and alter the reslr_mcmc function default iteration and MCMC settings. Also, it provides high-level summaries of the estimated parameters.

# S3 method for reslr_output
summary(object, ...)

Arguments

object

Output object from the reslr_mcmc

...

Not in use

Value

A list containing convergence diagnostics and parameter estimates for the output.

Examples

# \donttest{
data <- NAACproxydata %>% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
jags_output <- reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t")
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 208
#>    Unobserved stochastic nodes: 107
#>    Total graph size: 1003
#> 
#> Initializing model
#> 
summary(object = jags_output)# }
#> No convergence issues detected. 
#> # A tibble: 3 × 7
#>   variable    mean      sd     mad      q5     q95  rhat
#>   <chr>      <num>   <num>   <num>   <num>   <num> <num>
#> 1 alpha    -1.99   0.0163  0.0160  -2.02   -1.96    1.00
#> 2 beta      0.823  0.0125  0.0124   0.802   0.844   1.00
#> 3 sigma_y   0.0662 0.00961 0.00960  0.0508  0.0824  1.00